streets_vehicle_list: Explicit project CMake name
The CMakeLists.txt file used the ${PROJECT_NAME} throughout the file
instead of the actual name. This can make it harder to understand the
file contents, and it makes finding things harder too. This commit
replaces the ${PROJECT_NAME} variable with the actual project name.

A common argument for using PROJECT_NAME is that it defines it in a
single spot, and changes to the project name automatically propagate.
However, changes can just as easily be made with a find and replace
tool. Another argument is that is save typing, but I disagree with this
because you still have to type ${PROJECT_NAME}, which is not a big
saving depending on the project name and it is more awkward to type.

streets_utils/streets_vehicle_list/CMakeLists.txt
* replace ${PROJECT_NAME} with the actual project name