streets_spat: Make GTest find_package conditional
Because unit tests are disabled when the project is consumed by another,
it does not make sense to require GTest in all settings. Therefore, the
find_package command for GTest has been put in a conditional. This way,
the package is only required when unit tests are being built.

streets_utils/streets_signal_phase_and_timing/dependencies.cmake
* put GTest find_package in a conditional to only be required when tests
  are being built.