streets_vehicle_list: Make GTest conditional
GTest was previously required in all instances, but it is only actually
required when unit tests are being built. If this project is consumed by
another project, then by default tests will not be built. In this
scenario, there is an unnecessary requirement on GTest.

streets_utils/streets_vehicle_list/dependencies.cmake
* make GTest find_package conditioned on building tests